cherry-pick-votes: Remove doc changes now cherry-picked
dacd91898f turned out to have nothing applicable to maint-5.20.

The rest all applied with no or only minimal conflicts.

Also cherry-picked the following Porting/release_schedule.pod commits which
were not on the list, but were required for the smooth application of ones
that were (they're all required anyway if the document is to be up to date):
ea61af38ef, c9c069a80c, 53e6b8f769 and cf217b1287.