Use more universal function in tests where makes sense
By using a different function that is just as applicable, or more so, in
these cases, we can avoid having to require a file.

The functions that translate code points between character sets are
always available and just as appropriate for the uses in the three tests
t/re/fold_grind.t, t/re/reg_fold.t, and lib/unicore/TestProp.pl.  Change
to use it instead of the functions that take string inputs.  This is
slightly cleaner, as the file that contains it no longer has to be
'required'.  And in some of the cases the code point function is a
better choice, as there had to be translation between chr and ord
before.