Fix regex pattern dump of ANYOF nodes
An ANYOF node is used to handle bracketed character classes and Unicode
properties.  It is implemented in part as a bit vector of code points
that match for smallish ordinals, plus other means to specify larger
code points, or code points that match only under certain conditions.
For example, under /d, whether [\w] matches word characters in the upper
Latin1 range depends on whether the target string is in UTF-8 or not,
so the characters which are conditionally matched are saved outside
the bitmap.  For performance, there is a flag that gets set if all code
points too large for the bitmap match.  This is a common occurrence for
complemented classes, such as [^a-z].  Prior to this commit, if this
flag was set, those things that are conditionally matched were not
output when dumping the regex pattern (under the command line option
-Dr, for example).