regcomp.c: Fix bug in /[A-Z]/i
This also fixes /[a-z]/i.

When not under /i, these two ranges alone in a bracketed character class
can be optimized into qr/[[:upper:]]/a and qr/[[:lower:]]/a respectively.
This optimization saves space in the pattern (as no bitmap is needed),
and I think it executes faster.  But this optimization has to be
foregone under /i (unless /a is also present) because otherwise
certain non-ASCII characters such as the \N{KELVIN SIGN} don't match,
and they should.