avoid $a["foo" eq $x ? 0 : 1] warning
RT #123609

The compiling code that looks for OP_MULTIDEREF candidates was
unconditionally numifying the first constant of an array index expression,
before it had confirmed that the expression consisted solely of a const.
So in something like

    $a['foo' eq $x ? 0 : 1]

the 'foo' would be numified and give a spurious warning:

    Argument "foo" isn't numeric

This commit fixes it by skipping the OP_MULTIDEREF optimisation if the
array index const isn't SvIOK(). In theory this means that something like

    $r->["0"]

no longer gets optimised, but I think we can live with that :-)

It also means that the test for the const being SvROK is no longer
necessary.

(Finally, I moved the declaration of the iv var down a few scopes as it
wasn't being used elsewhere in the wider scope.)