regcomp.c: Refactor a calculation
Currently the way we calculate if the endpoints in a range in a
[bracketed character class] are "literal" (like 'A', 'b') vs non (like
\x{41}) is to have a count of the literal endpoints.

Future commits will expand the definition of literal to include things
that are portably-specified, including things like \t, \N{U+xx}, etc.
It will be easier to specify that we have encountered a non-portable
name instead of the other way around.  So that is what this commit does.
The only non-portables are \digit, \o{}, \x{}, and \cX for all X.