Add portablity warning for re 'strict'
When a range in a bracketed character class has one end be specified as
Unicode, the whole range is viewed as Unicode.  Currently this is not
warned about, though it is somewhat like mixing apples and oranges.
This commit adds a warning, but only under "use re 'strict'", and
it now documents the only one-end behavior.