Fix SDBM_File clean target after ac0e53cd2e.
The .dir extension is the extension of actual directory files in
the file system on VMS, so putting *.dir in the clean target has
us removing all the directories under ext/SDBM_file, including the
t/ directory.  Since the tests are in MANIFEST, a subsequent
reconfigure in the same build tree would then fail its manifest
check, such as when smoking multiple configurations.

So let's only clean up the files created by SDBM, which end with
.sdbm_dir on VMS.