attributes.xs: Don’t emit const warning for anons
The intent was to emit the warning for any cases other than custom
attributes on anonymous subs.  I did that by checking to see whether
we had a closure prototype, warning otherwise.  But anonymous subs
that are not closures are not closure prototypes.  So we need to
check CvANON instead.