regcomp.c: Minor optimizations
\d, [:digit:], and [:xdigit:] don't match anything in the upper Latin1
range.  Therefore whether or not the target string is UTF-8 or not
doesn't change what they match, hence the /d modifier acts exactly like
the /u modifier for them.  At run-time /u executes fewer branches
because it doesn't have to test if the target string is in UTF-8 or not,
so treating these as if /u had instead been specified saves some
runtime.