cherry-pick-votes: Drop the fixes for #122283 and #122890 from the list
Only one voter (arc) came forward for the #122890 fix, and that was on the
basis that only some of the series of commits involved be cherry-picked,
with a note that they were not likely to apply cleanly either...

This is all rather troubling, and hv agreed that it's too complex for
maint: http://www.nntp.perl.org/group/perl.perl5.porters/2015/01/msg225074.html

So #122890 must be dropped since it's controversial at best.

Therefore, #122283 must also be dropped since it's also controversial
without the fix for #122890, which #122283 caused: as hv noted, #122283 is
merely suboptimal memory usage, whereas #122890 actually causes wrong
results: http://www.nntp.perl.org/group/perl.perl5.porters/2015/01/msg225056.html

arc also expressed a preference for "both fixes or neither fix" since
releasing 5.20.2 with a bug that was known in 5.20.0 and 5.20.1 is less bad
than releasing it with a different known bug.