Revert the support for new warning categories outside of "all"
This reverts & amends my v5.21.7-151-gea5519d and Karl Williamson's
v5.21.7-183-g2f3cbe1, the latter was only need because of the
former.

I've also taken the opportunity to fix the long-standing trivial bug
with misaligned code in warnings.{pm,h}. That was easier to commit along
with this than to split it up from the other generated changes.

Why revert this? See the "use warnings 'absolutely-all-almost';" thread
on perl5-porters for the latest summary:

    http://www.nntp.perl.org/group/perl.perl5.porters/2015/01/msg225066.html

Basically as I explained in v5.21.7-151-gea5519d the current design of
the API makes it too contentious to freely add new warnings, but there's
no consensus on how to solve that. I.e. whether we should just add them
to "all", or do this change, or several other possible things outlined
in that thread and elsewhere.

Since the deadline for contentious changes for v5.22 is already past us
I'm backing this out for now.