Dodge warning with STATIC_ASSERT_STMT, VMS C++.
This is not a C++11 compiler so it's getting the emulated version
of static_assert and it doesn't like it in one case:

        STATIC_ASSERT_STMT(SVf_FAKE >= 1<<(sizeof(PadnameFLAGS(pn)) * 8));
........^
%CXX-W-REFNESTFUNVAR, reference to local variable of enclosing function is
          not allowed
at line number 3562 in file D0:[craig.blead.ext.B]b.c;1

While pn is in fact a local variable, it's in an enclosing struct,
not an enclosing function.  Nevertheless, we don't actually need
anything from pn other than the size of one of its members, so
we'll just use a NULL pointer of the correct type rather than a
real live instance of that type.  This is supposedly C89 so
shouldn't break anything else.