rename unop_aux_stringify to multideref_stringify
This function returns a string representation of the OP_MULTIDEREF op
(as used by the output of perl -Dt).

However, the stringification of a UNOP_AUX op is op-specific, and
hypothetical future UNOP_AUX-class ops will need their own functions.
So the current function name is misleading.

It should be safe to rename it, as it only been in since 5.21.7, and isn't
public.