cherry-pick-votes: Remove three cherry-picked commits
I've taken the executive decision to cherry-pick these harmless commits
without votes.

(I've also been updating the RMG (again, without votes) to get it exactly
in sync with blead except for things that don't apply to maint-5.20. This
removes the ambiguity of which RMG to follow when doing a MAINT release.)