"A" =~ '\N{U+41}' (partial) should be true on all platforms
The code for handling this only worked in double-quotish contexts.  To
make it work in single-quotish areas as well, it needs to be moved out
of toke.c, and corresponding code added to regcomp.c.  This commit does
just the portion that removes the code from toke.c.  The other portion
hasn't been fully debugged yet.  This means that blead will now fail on
EBCDIC platforms in double-quotish contexts.  But EBCDIC platforms
aren't fully supported in blead currently anyway.

The reason this partial commit is being pushed to blead now is that its
absence is blocking other work in toke.c

Spotted by Father Chrysostomos