toke.c: Simplify \N{U+...} code
If we are parsing a \N{U+XXX.YYY} construct in a regexp literal, we do
not need to pass it to grok_hex, because we do not need the numeric
value at this point.  The regexp engine will be calling grok_hex
again, after all.  A simple scan for hex digits should be faster, and
makes the code a little simpler, too.