Make nowait spawn return correct pid on VMS.
Apparently I missed this back in eed5d6a149b02.  When spawning a
subprocess without waiting, the return value is the pid of that
process, but we need to let pp_system know we're doing that so it
doesn't manipulate the value as if it were an exit status.

One symptom was that t/test.pl's watchdog would sometimes report
that it had failed to start when in fact it had started just fine.