regen/mk_invlists.pl: Add capability for line break props
This is a partial implementation of a full inversion map generation
capability, which is why some code is indented more than necessary --
in the future there will be things that use that.  But this is
sufficient for 5.22.

This allows for the generation of tables to handle the Unicode line
breaking properties, like GCB and WB.  Future commits will actually use
this capability.