regen/mk_invlists.pl: Revamp #if generation
This changes where the symbols are defined to a single file each.  This
may save text space, depending on the compiler.  The next commit will
cause this hdr to be included in more places, so it becomes more
important to do this.

At the same time this removes the guard for #ifndef PERL_IN_XSUB_RE.
The code now is executed regardless of that.  This is simpler, and
previously there might have been the possibility of uninitialized memory
being read, should re_comp.o be executed before recomp.o.