Fix some style issues in vms/vms.c.
This normalizes function definitions and addresses the most
egregious of the Totally Forgot to Indent At All problems that are
rampant throughout this file.

For now I've punted on normalizing indents and tried to be
consistent with whatever local variant seems to be in use. There
are 2-, 3- (!), and 4-space indents, along with a liberal
sprinkling of tab characters, most (but not all) of which seem to
be intended to be equivalent to 8 spaces.

I considered normalizing indents with some like

   astyle -c --style=linux vms.c

but that produces 8400+ changed lines, which doesn't seem worth it.