pp_pack.c: Silence compiler warning
This was introduced by 9df874cdaa2f196cc11fbd7b82a85690c243eb9f
in changing the name of some static functions.  I didn't realize at the
time that the function was defined in embed.fnc, as none of the others
are, and it was always called with the S_ prefix form.  Nor did I notice
the compiler warnings.

It turns out that the base name of this function is the same as a public
function, so I've renamed it to have prefix 'S_my_'.