init a var in sv_clear() to keep Coverity happy.
When a new hash is found to iterate over, the old value of hash_index (if
any) is temporarily saved (in the new hash), then restored after the hash
has been iterated over.
This means that the first hash encountered causes the random initial
value of hash_index to be saved and then restored. This is perfectly
harmless, but makes Coverity unhappy. Since other such tools might also
in future detect such dabbling with uninitialised values, its simplest
(but unnecessary) to initialise this var to zero.